skill_sphere_mcp/pyproject.toml
[project]
name = "skill-sphere-mcp"
version = "0.1.0"
description = "Management Control Plane for Skill Sphere"
authors = [
    {name = "Bernd Prager", email = "bernd@example.com"},
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "neo4j>=5.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "opentelemetry-instrumentation-requests>=0.42b0",
    "python-dotenv>=1.0.0",
    "scipy>=1.11.0,<1.14.0",  # Pin scipy to compatible version
    "scikit-learn>=1.3.0",
    "networkx>=3.4.2",
    "sentence-transformers>=4.1.0",
    "pytest>=8.3.5",
    "node2vec>=0.4.3",
    "numpy>=1.24.0",
    "gensim<4.3.0",
    "httpx>=0.28.1",
    "markitdown>=0.1.1",
    "pylint>=3.3.7",
    "pytest-cov>=6.1.1",
    "pytest-asyncio>=0.21.0",
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "ruff>=0.1.0",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest>=7.4.0",
    "ruff>=0.1.0",
    "httpx>=0.28.1",
    "markitdown>=0.1.1",
    "pylint>=3.3.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = ["E501", "B002"]

[tool.ruff.isort]
known-first-party = ["skill_sphere_mcp", "hypergraph"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=skill_sphere_mcp"
pythonpath = ["src"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = ["numpy.*", "sklearn.*", "gensim.*", "node2vec.*"]
ignore_missing_imports = true

[tool.coverage.run]
omit = ["hypergraph/skills_hypergraph_ontology.py"]
